requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
python-dotenv
dj-database-url
psycopg2-binary
gunicorn
whitenoise
